feat: reorg language files and add fr-CA
