+++ src/translate/es-main.ts
@@ -1,0 +1,116 @@
+// prettier-ignore
+const ES = {
+  "Address has been copied to the clipboard": "La dirección se ha copiado al portapapeles",
+  "Amount": undefined,
+  "An account with the same email exists already": undefined,
+  "Cancel": "Cancelar",
+  "Click QR code to copy": "Toque el código QR para copiar",
+  "Close": "Cerrar",
+  "Color Theme": undefined,
+  "Confirm": "Confirmar",
+  "Contacts": "Contactos",
+  "Create Account": undefined,
+  "Create new account": undefined,
+  "Current Balance": "Saldo actual",
+  "Dark mode": undefined,
+  "Default (OS)": "Predeterminado (OS)",
+  "Description": "Descripción",
+  "Email Support": "Soporte de contacto",
+  "Email": "Correo electrónico",
+  "English": "Inglés",
+  "Enter amount and destination": "Ingrese cantidad y dirección",
+  "Enter amount": "Ingrese cantidad",
+  "Enter your phone number and we'll text you an access code": "Ingrese su número de teléfono y le enviaremos un mensaje de texto con un código de acceso",
+  "Fee": "Tarifa",
+  "First Name": undefined,
+  "From": "Desde",
+  "Generate new invoice for %{amount}": "Generar una factura nueva por %{amount}",
+  "Generate New Invoice": "Generar una factura nueva",
+  "Hash": undefined,
+  "History": "Historial",
+  "Home": "Inicio",
+  "How much would you like to send?": undefined,
+  "Invaild email or password": undefined,
+  "Invalid QR Code": undefined,
+  "Invalid verification. Please try again": undefined,
+  "Invite": undefined,
+  "Invoice Expired...": undefined,
+  "Invoice has been copied to the clipboard": undefined,
+  "Invoice is already paid": undefined,
+  "Invoice value is now %{value}": undefined,
+  "Language": "Idioma",
+  "Last Name": undefined,
+  "Light mode": undefined,
+  "Lightning": undefined,
+  "Link has been copied to the clipboard": undefined,
+  "Load more transactions": "Obtener más transacciones",
+  "Login to send": undefined,
+  "Login": "Iniciar sesión",
+  "Logout": "Cerrar sesión",
+  "New Password": undefined,
+  "Next": "Próximo",
+  "No Contacts": undefined,
+  "No transactions": "No transacciones",
+  "Note or Label": undefined,
+  "Note": undefined,
+  "OnChain": undefined,
+  "Password is too short": undefined,
+  "Password": undefined,
+  "Payment amount exceeds balance of %{balance} sats": "Cantidad excede su saldo de por %{balance}",
+  "Payment failed": undefined,
+  "Phone number is too long": undefined,
+  "Phone number is too short": undefined,
+  "Please try again later.": undefined,
+  "Please try again later": undefined,
+  "Portuguese": "Portugués",
+  "Powered By": undefined,
+  "Previous months": "Meses anteriores",
+  "Receive another payment": undefined,
+  "Receive Bitcoin": "Recibir Bitcoin",
+  "Receive": "Recibir",
+  "Received from": "Recibido de",
+  "Recent Transactions": "Las transacciones",
+  "Recover Account": undefined,
+  "Recover your account": undefined,
+  "Refresh": undefined,
+  "Reload Application": undefined,
+  "Save": undefined,
+  "Scan QR code": "Escanear QR",
+  "Send another payment": undefined,
+  "Send Bitcoin": "Enviar Bitcoin",
+  "Send Payment": "Enviar Pago",
+  "Send": "Enviar",
+  "Sending bitcoin to": undefined,
+  "Sent to": "Enviado a",
+  "Set a note for the receiver here (optional)": undefined,
+  "Set a note for the sender here (optional)": undefined,
+  "Set Amount": undefined,
+  "Set invoice value in %{currency}": undefined,
+  "Set value to send in %{currency}": undefined,
+  "Settings": "Configuración",
+  "Share Bitcoin Address": "Compartir la dirección de Bitcoin",
+  "Share Lightning Address": "Compartir la dirección de Lightning",
+  "Sorry, we appear to be having issues loading the application data. If the problem persists, please contact support.": undefined,
+  "Spanish": "Español",
+  "This month": "Este mes",
+  "To confirm your phone number, enter the code we just sent you on %{phoneNumber}": "Ingrese su número de teléfono, y le enviaremos un mensaje de texto con un código de acceso.",
+  "To": "Enviado a",
+  "Today": "Hoy",
+  "Transaction List": undefined,
+  "Transactions with %{contactUsername}": "Las transacciones con %{contactUsername}",
+  "username, invoice, or address": undefined,
+  "Username": "Nombre de usuario",
+  "Verify Account": undefined,
+  "Verify you are human": undefined,
+  "Waiting for payment confirmation...": undefined,
+  "What for": undefined,
+  "Yesterday": "Ayer",
+  "You can't send to yourself": undefined,
+  "You successfully recovered your account. Please change your password.": undefined,
+  "Your account verification is still pending": undefined,
+  "Your email address has been verified": undefined,
+} as const
+
+export type TranslationKey = keyof typeof ES
+
+export default ES
